+++ sstring_cat.c
@@ -1,6 +1,6 @@
 #include "sstring.h"
 
-unsigned int sstring_cat(sstring *ds, const sstring *ss)
+unsigned long sstring_cat(sstring *ds, const sstring *ss)
 {
   return sstring_catb(ds, ss->s, ss->len);
 }
