+++ sstring_copy.c
@@ -1,6 +1,6 @@
 #include "sstring.h"
 
-unsigned int sstring_copy(sstring *ds, const sstring *ss)
+unsigned long sstring_copy(sstring *ds, const sstring *ss)
 {
   return sstring_cpyb(ds, ss->s, ss->a);
 }
